+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2025-12-02
+- Added "config" to announcements helper command
+- Cleaned up announcements helper command
+
 ## 2025-12-01
 - Added announcements helper command
 - Added motd hint